requirements_deploy.txt
# Core dependencies for deployment
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Document processing
pdfminer.six>=20221105
python-docx>=0.8.11
beautifulsoup4>=4.12.0
lxml>=4.9.0

# NLP and ML (optimized versions)
spacy>=3.7.0
transformers>=4.35.0
torch>=2.0.0
scikit-learn>=1.3.0

# Template engine
jinja2>=3.1.0

# Additional utilities
chardet>=5.2.0
tqdm>=4.66.0

# For deployment
click>=8.0.0
